% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagehistory.R
\name{action}
\alias{action}
\title{An action represents the changes that have
been done to a certain file. Currently, it
contains brightness, contrast, gamma, and
crop information.}
\usage{
action(bright, cont, gam, crops)
}
\arguments{
\item{bright}{the current brightness value.}

\item{cont}{the current contrast value}

\item{gam}{the gamma value}

\item{crops}{the coordinates of crops}
}
\value{
A concatenated sequence of action values.
}
\description{
An action represents the changes that have
been done to a certain file. Currently, it
contains brightness, contrast, gamma, and
crop information.
}
\examples{
Should not be used by the user.

}
