% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny-function}
\alias{runShiny-function}
\alias{runShiny}
\title{runShiny function to start shiny app}
\usage{
runShiny(current_shinyimg)
}
\description{
A function to start shiny app
}
\section{Fields}{

\describe{
\item{\code{current_shinyimg}}{takes in an optional parameter of a image user is
currently editing in the command line}
}}

