% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagehistory.R
\docType{class}
\name{shinygui}
\alias{shinygui}
\title{The ShinyImg GUI editor}
\format{\code{\link{R6Class}} object.}
\usage{
shinygui
}
\value{
Object of \code{\link{R6Class}} with manipulation functions.
}
\description{
The ShinyImg GUI editor
}
\examples{
local_tiger = shinyimg$new("Tigerwater_edit2.jpg")
editor_instance = shinygui$new()
editor_instance$load(tiger)
#' @section Methods:
\\describe{
  \\item{Documentation}{The user should create an instance to facilitate the gui editing of an image.}
  \\item{\\code{new()}}{Default constructor. No arguments required.}

  \\item{\\code{load()}}{Loads a ShinyImage in order to process.}
  }

}
\keyword{data}
