% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagehistory.R
\name{shinyimg}
\alias{shinyimg}
\title{An EBImage wrapper with integrated history tracking.}
\usage{
shinyimg(inputImage)
}
\arguments{
\item{inputImage}{A URL or the local path of an image}
}
\value{
A ShinyImg object with manipulation functions

a image that can be used in Shiny
}
\description{
An EBImage wrapper with integrated history tracking.
}
\examples{
local_tiger = shinyimg("Tigerwater_edit2.jpg")
web_tiger = shinyimg("https://upload.wikimedia.org/wikipedia/commons/1/1c/Tigerwater_edit2.jpg")

Undoes the last change done to this image. 
When the original image state is reached,
no more undos are possible.
local_tiger = shinyimg("Tigerwater_edit2.jpg")
local_tiger$add_brightness()
local_tiger$undo() # Undoes the brightness addition
local_tiger$redo() # Redoes the brightness addition
Returns a Shiny compatible image
local_tiger = shinyimg("Tigerwater_edit2.jpg")
local_tiger$add_brightness()
shiny_tiger = local_tiger$getimg() #Now usable by Shiny

Adds brightness to the image
local_tiger = shinyimg("Tigerwater_edit2.jpg")
local_tiger$add_brightness()
Removes brightness (darkens) to the image
local_tiger = shinyimg("Tigerwater_edit2.jpg")
local_tiger$remove_brightness()
Adds contrast to the image
local_tiger = shinyimg("Tigerwater_edit2.jpg")
local_tiger$add_contrast()
Removes contrast from the image
local_tiger = shinyimg("Tigerwater_edit2.jpg")
local_tiger$remove_contrast()
After two points are selected, a cropping selection can be create
in order to crop the image to the desired size.
local_tiger = shinyimg("Tigerwater_edit2.jpg")
local_tiger$crop()
}
