% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagehistory.R
\name{shinyload}
\alias{shinyload}
\title{Wrapper to load an image from a cold boot.}
\usage{
shinyload(filename)
}
\arguments{
\item{filename}{The filename of a file previously generated by shinyimg's $save function.}
}
\description{
Wrapper to load an image from a cold boot.
}
\examples{
web_tiger = shinyimg$new(system.file("images","sample.jpg",package="ShinyImage"))
web_tiger$save("tiger.si")
# Restart R 
reloaded_tiger = shinyload("tiger.si")
}
