% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagehistory.R
\name{start_gui}
\alias{start_gui}
\title{A Shiny application that requires a ShinyImg to edit
(most likely will be changed in the future)}
\usage{
start_gui(img)
}
\arguments{
\item{img}{A ShinyImg to edit using the GUI editor}
}
\description{
A Shiny application that requires a ShinyImg to edit
(most likely will be changed in the future)
}
\examples{
local_tiger = shinyimg("Tigerwater_edit2.jpg")
start_gui(local_tiger)

}
