% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny-function}
\alias{runShiny-function}
\alias{runShiny}
\title{runShiny function to start shiny app}
\usage{
runShiny(current_shinyimg)
}
\arguments{
\item{current_shinyimg}{takes in an optional parameter of a image user is
currently editing in the command line}
}
\description{
A function to start shiny app
}
\examples{
\dontrun{
   runShiny()
}
}
