% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagehistory.R
\docType{class}
\name{shinyimg}
\alias{shinyimg}
\title{An EBImage wrapper with integrated history tracking.}
\format{\code{\link{R6Class}} object.}
\usage{
shinyimg
}
\value{
Object of \code{\link{R6Class}} with manipulation functions.
}
\description{
An EBImage wrapper with integrated history tracking.
}
\section{Methods}{

\describe{
  \item{Documentation}{The user should not need to create an action object. This is a class used exclusively by a shinyimg to keep track of a set of changes.}
  \item{\code{new(img)}}{Default constructor. \code{img} can be either a URL or a location of a local image.}

  \item{\code{undo()}}{Undoes the last change done to this image. When the original image state is reached, no more undos are possible.}
  \item{\code{redo()}}{Redos the next action after an undo has been performed. Will no longer redo if there are no more undos to redo.}
  \item{\code{getimg()}}{Returns a Shiny compatible image.}
  \item{\code{add_brightness()}}{Adds brightness to the image.}
  \item{\code{remove_brightness()}}{Removes brightness (darkens) to the image.}
  \item{\code{add_contrast()}}{Adds contrast to the image.}
  \item{\code{remove_contrast()}}{Removes contrast from the image.}
  \item{\code{crop()}}{Uses locator to get corners of an image. Automatically finds min and max coordinates. After two points are selected, a cropping selection can be create in order to crop the image to the desired size.}
  \item{\code{save(filepath)}}{Saves the current state to be resumed later. \code{filepath} has a default value of 'workspace.si'}
  \item{\code{load(filepath)}}{Loads a previously saved state. \code{filepath} has a default value of 'workspace.si'}
  \item{\code{dim()}}{Returns the current image dimentions.}
  \item{\code{render()}}{Renders the current image.}
  \item{\code{toggle_render()}}{Toggles the automatic rendering after making a change. By default, this option is off.}
  }
}

\examples{


web_tiger = shinyimg$new('https://upload.wikimedia.org/wikipedia/commons/1/1c/Tigerwater_edit2.jpg')

web_tiger$add_brightness()
web_tiger$undo() # Undoes the brightness addition

web_tiger$redo() # Redoes the brightness addition

shiny_tiger = web_tiger$getimg() # Now usable by Shiny

web_tiger$add_brightness() # Adds brightness to the image

web_tiger$remove_brightness() # Removes brightness

web_tiger$add_contrast() # Adds contrast

web_tiger$remove_contrast() # Removes Contrast

web_tiger$crop() # Allows the user to select two points to crop to.

web_tiger$save('save.ri') # Saves the current state. The filename is optional.

web_tiger$load('save.ri') # Loads from a previously saved state. The filename is optional. Requires a previously instantiated shinyimg instance (argument provided to new can be null).

}
\keyword{data}
